requirements.txt
playwright==1.40.0
openai==1.3.0
python-dotenv==1.0.0
beautifulsoup4==4.12.2
lxml==4.9.3
google-generativeai==0.3.0

